Add initial web app files including HTML, manifest, and service worker
- Created index.html for the main web interface with service worker registration.
- Added manifest.json for PWA configuration with app details and icons.
- Implemented service-worker.js to handle caching of assets and version management.
