+++ .gitignore
@@ -1,5 +1,4 @@
 node_modules/
 processing/crossrefs*
 processing/topics*
-dest/main.css.map
-dest/main.js.map
+dest/*
