--- dest/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="main.css" />
-    <title>Touch Grass Bible</title>
-  </head>
-  <body>
-    <script src="main.js"></script>
-  </body>
-</html>
