--- dest/main.css
@@ -1,6 +1,0 @@
-/*
-THIS IS A GENERATED/BUNDLED FILE BY ESBUILD
-if you want to view the source, please visit the github repository of this app
-*/
-
-body{margin:0;padding:0;user-select:none}.AppShellElement{width:100vw;height:100vh;overflow:hidden;display:flex;flex-direction:column;background-color:#000;color:#fff;font-family:Arial,sans-serif;font-size:16px}input,button,textarea,select,.input{border-radius:.6em .2em}.command-palette{position:fixed;top:50%;left:50%;transform:translate(-50%,-50%);width:calc(100% - 2em);height:calc(100% - 2em);max-width:var(--readingwidth);background:#0008;border-radius:8px;box-shadow:0 0 16px -8px #fff;backdrop-filter:blur(8px);z-index:9999;font-family:sans-serif;border:1px solid #fff6;user-select:none}.palette{display:flex;flex-direction:column;padding:8px;height:100%;box-sizing:border-box;color:var(--foreground);background-color:transparent}.palette-header{display:flex;align-items:center;justify-content:space-between;padding:8px 12px;background-color:#fff1;border-radius:4px;color:var(--foreground)}.palette-header button.active{background-color:var(--accent1);color:#000}.palette-search{padding:10px 14px;border:none;border-radius:4px;background-color:#fff1;color:var(--foreground);font-size:1rem;outline:none;transition:background-color .2s,box-shadow .2s;width:100%;box-sizing:border-box}.palette-search::placeholder{color:#888}.palette-search:focus{background-color:#fff2}.palette-content{margin-top:8px;overflow-y:auto;flex:1}.palette-content .category-title{padding:12px 12px 8px;color:#fff8;text-align:end;cursor:pointer}.palette-content .category-title:hover{color:#fffa}.command-item{border-top:1px solid #fff4;cursor:pointer;transition:background-color .2s,transform .1s;display:flex;align-items:center;color:var(--foreground)}.command-item.selected{background-color:#fff2}.command-item-info{padding:8px 12px;flex:1 1 auto}.command-item .command-description{color:#fff5}.command-item .command-description.hidden{max-height:0;overflow:hidden}.command-item.selected .command-description.hidden{max-height:500px}.command-item .command-subsearch{padding:0 2em;color:#fff5;align-self:stretch;display:flex;align-items:center}.command-item .command-subsearch:hover{color:#fffa}.navbar{display:flex;justify-content:space-between;align-items:center;padding:.5em;background-color:#fff2;cursor:pointer;color:var(--foreground);border-bottom:1px solid #fff4;z-index:1}.navBarTitle{font-size:1.5rem;font-family:var(--font)}.content{padding:2em;background-color:var(--background);color:var(--foreground);font-family:var(--font);font-size:var(--font-size);overflow:scroll}button{align-items:center;background-color:#fff1;border:0;border-radius:.5em 1em;color:#fff;cursor:pointer;display:inline-flex;justify-content:center;outline:none;padding:.5em 1em;white-space:nowrap;user-select:none}button:hover{background-color:#fff2}body{--background: #000f;--foreground: #ffff;--readingwidth: 25em;--navbarheight: 5em;--textdisplay: block}.highlighted-query{color:#f0fc}.content{.verse{text-indent:1em;padding:.4em;max-width:var(--readingwidth);margin:auto;display:var(--textdisplay);cursor:pointer}.verse:hover{text-shadow:0 0 5px #fff4}.verse .verseNumber{color:#fff8;padding:.5em}.verseActive .verseNumber{color:#f0fc}.verse .versePBreak{color:#fff8}>:last-child{min-height:100%;padding-bottom:2rem}}.content.wrap{.verse{display:inline}div.versePBreak:before{display:block;content:""}.versePBreak .verseNumber,>:first-child .verseNumber{font-size:2em;text-transform:capitalize;float:left;line-height:.8;padding:.2em .1em 0 0}.versePBreak .versePBreak{display:none}}
