--- dest/main.js
@@ -1,24 +1,0 @@
-/*
-THIS IS A GENERATED/BUNDLED FILE BY ESBUILD
-if you want to view the source, please visit the github repository of this app
-*/
-
-var X=Object.defineProperty;var W=(l,t,e)=>t in l?X(l,t,{enumerable:!0,configurable:!0,writable:!0,value:e}):l[t]=e;var r=(l,t,e)=>W(l,typeof t!="symbol"?t+"":t,e);var U={xmlns:"http://www.w3.org/2000/svg",width:24,height:24,viewBox:"0 0 24 24",fill:"none",stroke:"currentColor","stroke-width":2,"stroke-linecap":"round","stroke-linejoin":"round"};var N=([l,t,e])=>{let a=document.createElementNS("http://www.w3.org/2000/svg",l);return Object.keys(t).forEach(o=>{a.setAttribute(o,String(t[o]))}),e?.length&&e.forEach(o=>{let f=N(o);a.appendChild(f)}),a},P=(l,t={})=>{let e="svg",a={...U,...t};return N([e,a,l])};var A=[["path",{d:"M2 16V4a2 2 0 0 1 2-2h11"}],["path",{d:"M22 18H11a2 2 0 1 0 0 4h10.5a.5.5 0 0 0 .5-.5v-15a.5.5 0 0 0-.5-.5H11a2 2 0 0 0-2 2v12"}],["path",{d:"M5 14H4a2 2 0 1 0 0 4h1"}]];var E=[["path",{d:"m9 18 6-6-6-6"}]];var k=[["path",{d:"m16 6 4 14"}],["path",{d:"M12 6v14"}],["path",{d:"M8 8v12"}],["path",{d:"M4 4v16"}]];var M=[["path",{d:"M3 12h.01"}],["path",{d:"M3 18h.01"}],["path",{d:"M3 6h.01"}],["path",{d:"M8 12h13"}],["path",{d:"M8 18h13"}],["path",{d:"M8 6h13"}]];var L=[["path",{d:"m21 21-4.34-4.34"}],["circle",{cx:"11",cy:"11",r:"8"}]];var B=[["path",{d:"M18 6 6 18"}],["path",{d:"m6 6 12 12"}]];var x=class{constructor(t,e){this.app=t;this.palette=e;r(this,"_commands",[]);r(this,"highlighter");r(this,"hili");r(this,"query")}setUp(t){return this.highlighter=new c([{regEXP:new RegExp(`(${t||"this will never match"})`,"ig"),cls:"highlighted-query"}]),this.hili=this.highlighter.highlight.bind(this.highlighter),this.query=t,this}tryTrigger(t){try{this.onTrigger(t)}catch(e){this.app.console.error(`Error in ${this.constructor.name}.onTrigger`,e)}return this}trygetCommands(t){try{return this.getCommands(t)}catch(e){return this.app.console.error(`Error in ${this.constructor.name}.getCommands`,e),[]}}tryrender(t,e){try{this.renderCommand(t,e)}catch(a){this.app.console.error(`Error in ${this.constructor.name}.renderCommand`,a)}return this}tryexecute(t){try{this.executeCommand(t)}catch(e){this.app.console.error(`Error in ${this.constructor.name}.executeCommand`,e)}return this}getcompatible(t,e,...a){if(!t)return e;let o=t.toLowerCase(),f=[];return a.map(d=>e.filter((u,i)=>f[i]?!1:(f[i]=d(u).toLowerCase().includes(o),f[i]))).flatMap(d=>d)}},T=class{constructor(t,e,a){this.command=e;this.palletCat=a;r(this,"el");r(this,"infoEl");r(this,"titleEl");r(this,"descriptionEl");r(this,"subEl");r(this,"hasSubsearch",!1);t.createEl("div",{cls:"command-item"},o=>{this.el=o,this.infoEl=o.createEl("div",{cls:"command-item-info"},f=>{this.titleEl=f.createEl("div",{cls:"command-title"}),this.descriptionEl=f.createEl("div",{cls:["command-description","hidden"]})}),this.subEl=o.createEl("div",{cls:"command-subsearch"},f=>{f.style.display=this.hasSubsearch?"flex":"none",f.setIcon(E)})})}setSubsearch(t){return this.hasSubsearch=t,this.subEl.style.display=t?"flex":"none",this}setTitle(t){return this.titleEl.replaceChildren(typeof t=="string"?this.palletCat.hili(t):t),this}setDescription(t){return this.descriptionEl.replaceChildren(typeof t=="string"?this.palletCat.hili(t):t),this}setHidden(t){this.descriptionEl.classList.toggle("hidden",t)}onClick(t){return this.el.addEventListener("click",e=>{e.stopPropagation(),t(e)}),this}onMouseEnter(t){return this.el.addEventListener("mousemove",t),this}onTouchStart(t){return this.el.addEventListener("touchstart",t),this}onSubClick(t){return this.subEl?.addEventListener("click",e=>{e.stopPropagation(),t(e)}),this}},S=class{constructor(t){this.app=t;r(this,"categories",[]);r(this,"containerEl",null);r(this,"paletteEl");r(this,"searchInputEl");r(this,"contentEl");r(this,"commandItems",[]);r(this,"selectedIndex",-1);r(this,"selectedEl");r(this,"currentCategory",null);r(this,"contexts",[]);r(this,"defaultContext");r(this,"inputMode","search");r(this,"headerEl");r(this,"maxResults",100);r(this,"topButtons",[]);r(this,"isOpen",!1);r(this,"handleOutsideClick",t=>{console.log("Outside click detected",t,this.containerEl,t.target,this.containerEl&&!this.containerEl.contains(t.target)&&!this.contentEl.contains(t.target)),this.containerEl&&!this.containerEl.contains(t.target)?this.close():this.searchInputEl.focus()});this.app.console.log("CommandPalette initialized")}addPalette(t){let e=typeof t=="function"?new t(this.app,this):t;return this.categories.push(e),this}open(t){this.contexts=[],this.defaultContext=t,this.isOpen=!0,this.display(t)}display(t){this.app.historyPush({name:"Command Palette",data:this.commandItems[this.selectedIndex]?.command}),this.contexts.push(t),this.inputMode="search",this.checkclose(),this.categories.forEach(e=>e.tryTrigger(t)),this.containerEl=this.app.contentEl.createEl("div",{cls:"command-palette"}),this.paletteEl=this.containerEl.createEl("div",{cls:"palette"}),this.headerEl=this.paletteEl.createEl("div",{cls:"palette-header"}),this.categories.forEach((e,a)=>{this.topButtons[a]=this.headerEl.createEl("button",{},o=>{o.setIcon(e.icon),o.addEventListener("click",()=>{this.toggleCategory(e),this.headerEl.querySelectorAll("button").forEach(f=>{f.classList.toggle("active",f===o&&this.currentCategory===e)})})})}),this.headerEl.createEl("button",{cls:"palette-back"},e=>(e.addEventListener("click",this.close.bind(this)),e.setIcon(B))),this.searchInputEl=this.paletteEl.createEl("input",{placeholder:"Search commands...",type:"search",cls:"palette-search"},e=>{e.inputMode=this.inputMode,e.addEventListener("input",()=>{this.update(e.value)}),e.addEventListener("keydown",a=>{switch(a.stopPropagation(),(a.metaKey?"Meta+":"")+(a.ctrlKey?"Ctrl+":"")+(a.altKey?"Alt+":"")+(a.shiftKey?"Shift+":"")+a.key){case"ArrowDown":a.preventDefault(),this.moveSelection(1);break;case"ArrowUp":a.preventDefault(),this.moveSelection(-1);break;case"Enter":a.preventDefault(),this.activateSelected();break;case"Escape":a.preventDefault(),this.close();break;case"ArrowRight":case"Tab":a.preventDefault(),this.gosubmenu(this.commandItems[this.selectedIndex]);break;case"ArrowLeft":case"Shift+Tab":a.preventDefault(),this.contexts.pop(),this.display(this.contexts.pop()||this.defaultContext);break}})}),this.contentEl=this.paletteEl.createEl("div",{cls:"palette-content"}),this.update(""),this.searchInputEl.focus(),document.addEventListener("click",this.handleOutsideClick)}gosubmenu(t){this.setCategory(t.palletCat.childtype),this.display(t.command)}handleBack(){this.contexts.length>1?(this.contexts.pop(),this.display(this.contexts.pop()||this.defaultContext)):this.close()}toggleCategory(t){this.currentCategory=this.currentCategory===t?null:t,this.update(this.searchInputEl.value)}setCategory(t){this.currentCategory=this.categories.find(e=>e.constructor.name===t)||null}checkclose(){this.containerEl&&(this.containerEl.remove(),this.containerEl=null,document.removeEventListener("click",this.handleOutsideClick))}close(){this.containerEl&&(this.containerEl.remove(),this.containerEl=null,document.removeEventListener("click",this.handleOutsideClick)),this.contexts=[],this.isOpen=!1}update(t){if(!this.containerEl)return;this.contentEl.empty(),this.commandItems=[],this.selectedIndex=0,(this.currentCategory?[this.currentCategory,...this.categories.filter(a=>a!==this.currentCategory)]:this.categories).forEach((a,o)=>{if(this.commandItems.length>this.maxResults)return;a.setUp(t);let f=a.trygetCommands(t),d=f.length!==0;d&&this.contentEl.createEl("div",{text:a.name,cls:"category-title"},u=>u.addEventListener("click",()=>this.toggleCategory(a))),this.topButtons[o].style.display=d?"flex":"none";for(let u of f){if(this.commandItems.length>this.maxResults)return;let i=this.commandItems.length,h=new T(this.contentEl,u,a).onClick(()=>a.tryexecute(u)).onMouseEnter(()=>this.selectIndex(i)).onTouchStart(()=>this.selectIndex(i)).onSubClick(()=>this.gosubmenu(this.commandItems[i]));a.tryrender(u,h),this.commandItems.push(h)}}),this.updateSelection()}moveSelection(t){let e=this.commandItems.length-1;this.selectIndex(Math.min(Math.max(this.selectedIndex+t,0),e))}selectIndex(t){this.selectedIndex=t,this.updateSelection()}updateSelection(){this.commandItems.forEach((t,e)=>{t.el.classList.toggle("selected",e===this.selectedIndex),e===this.selectedIndex&&(t.el.scrollIntoView({block:"nearest",inline:"nearest",behavior:"smooth"}),this.selectedEl=t.el)})}activateSelected(){this.selectedIndex>=0&&this.selectedIndex<this.commandItems.length&&this.selectedEl.click()}};var c=class{constructor(t){r(this,"args");this.args=t}highlight(t){let e=[];for(let f of this.args){let d=f.regEXP,u;if(d.global)for(;(u=d.exec(t))!==null;)e.push({start:u.index,end:u.index+u[0].length,type:f,matchText:u[0]}),u[0].length===0&&d.lastIndex++;else(u=d.exec(t))!==null&&e.push({start:u.index,end:u.index+u[0].length,type:f,matchText:u[0]})}e.sort((f,d)=>f.start-d.start);let a=document.createDocumentFragment(),o=0;for(let f of e){f.start>o&&a.appendChild(document.createTextNode(t.substring(o,f.start)));let d=document.createElement(f.type.elTag||"span");f.type.cls&&(d.className=f.type.cls);let u=f.matchText.replace(f.type.regEXP,f.type.replace||"$1");d.textContent=u,a.appendChild(d),o=f.end}return o<t.length&&a.appendChild(document.createTextNode(t.substring(o))),a}};var w=class{constructor(t,e){r(this,"enabled");r(this,"consoleName");this.enabled=t,this.consoleName=e}log(...t){this.enabled&&console.log(this.consoleName,...t)}info(...t){this.enabled&&console.info(this.consoleName,...t)}warn(...t){this.enabled&&console.warn(this.consoleName,...t)}error(...t){this.enabled&&console.error(this.consoleName,...t)}header(...t){this.enabled&&console.log("%c"+this.consoleName,...t)}};Node.prototype.addClass=function(...l){return this.classList.add(...l),this};Node.prototype.createEl=function(l,t,e){let a=document.createElement(l);if(t)if(typeof t=="string")a.textContent=t;else{let o=t;if(o.cls!==void 0&&(Array.isArray(o.cls)?a.className=o.cls.join(" "):a.className=o.cls),o.text!==void 0&&(typeof o.text=="string"?a.textContent=o.text:o.text instanceof DocumentFragment&&a.appendChild(o.text)),o.attr!==void 0)for(let[f,d]of Object.entries(o.attr))d==null||d===!1?a.removeAttribute(f):typeof d=="boolean"?d?a.setAttribute(f,""):a.removeAttribute(f):a.setAttribute(f,String(d));o.title!==void 0&&a.setAttribute("title",o.title),o.value!==void 0&&(a.value=o.value),o.type!==void 0&&(a.type=o.type),o.placeholder!==void 0&&(a.placeholder=o.placeholder),o.href!==void 0&&"href"in a&&(a.href=o.href)}return e&&e(a),this.appendChild(a),a};Node.prototype.empty=function(){return this.textContent="",this};HTMLElement.prototype.setIcon=function(l){return this.appendChild(P(l,{"stroke-width":1})),this};String.prototype.toTitleCase=function(){return this.toLowerCase().split(" ").map(l=>l.charAt(0).toUpperCase()+l.slice(1)).join(" ")};var b=class{constructor(t,e){this.app=e;r(this,"header");r(this,"content");r(this,"titleEl");this.header=t.createEl("div",{cls:"navbar"},a=>{this.titleEl=a.createEl("div",{cls:"navBarTitle",text:"Touch Grass Bible"}),a.addEventListener("click",this.ontitleclick.bind(this))}),this.content=t.createEl("div",{cls:"content"})}get title(){return this.app.title}onmenuclick(t){return this.header.addEventListener("contextmenu",t,!0),this}set title(t){this.app.title=t,this.titleEl.textContent=t}update(){this.content.empty()}ontitleclick(t){return this.header.addEventListener("click",t),this}};var y=class{constructor(t,e){this.doc=t;this._title=e;r(this,"console");r(this,"contentEl");r(this,"historyStack",[]);this.console=new w(!0,`${this._title||"App"}:`),this.contentEl=this.doc.body.createEl("div",{cls:"AppShellElement"}),this.title=this._title,document.addEventListener("DOMContentLoaded",()=>this.onload()),window.addEventListener("beforeunload",a=>{this.exit()||(a.preventDefault(),a.returnValue="")}),window.addEventListener("popstate",this.handlePopState.bind(this))}exit(){return this.unload()}load(){this.onload()}unload(){return this.onunload?.()||!0}historyPush(t){let e={name:t.name??"",time:new Date,data:t.data??null};this.historyStack.push(e),history.pushState(e,"","")}historyPop(){if(this.historyStack.pop(),this.historyStack.length===0){this.exit();return}this.onHistoryPop(this.historyStack[this.historyStack.length-1])}handlePopState(){this.historyPop()}async onload(){}async saveData(t){localStorage.setItem("app-data",JSON.stringify(t))}async loadData(){let t=localStorage.getItem("app-data");return Promise.resolve(t?JSON.parse(t):{})}get title(){return this.doc.title}set title(t){this.doc.title=t||this._title}async loadJSON(t){try{return console.log(`${t} loaded`),await(await fetch(t)).json()}catch(e){return console.error(`Failed to load ${t}:`,e),Promise.reject(e)}}};var V=["GENESIS","EXODUS","LEVITICUS","NUMBERS","DEUTERONOMY","JOSHUA","JUDGES","RUTH","1 SAMUEL","2 SAMUEL","1 KINGS","2 KINGS","1 CHRONICLES","2 CHRONICLES","EZRA","NEHEMIAH","ESTHER","JOB","PSALMS","PROVERBS","ECCLESIASTES","SONG SOLOMON","ISAIAH","JEREMIAH","LAMENTATIONS","EZEKIEL","DANIEL","HOSEA","JOEL","AMOS","OBADIAH","JONAH","MICAH","NAHUM","HABAKKUK","ZEPHANIAH","HAGGAI","ZECHARIAH","MALACHI","MATTHEW","MARK","LUKE","JOHN","ACTS","ROMANS","1 CORINTHIANS","2 CORINTHIANS","GALATIANS","EPHESIANS","PHILIPPIANS","COLOSSIANS","1 THESSALONIANS","2 THESSALONIANS","1 TIMOTHY","2 TIMOTHY","TITUS","PHILEMON","HEBREWS","JAMES","1 PETER","2 PETER","1 JOHN","2 JOHN","3 JOHN","JUDE","REVELATION"],G=["Gen","Exod","Lev","Num","Deut","Josh","Judg","Ruth","1Sam","2Sam","1Kgs","2Kgs","1Chr","2Chr","Ezra","Neh","Esth","Job","Ps","Prov","Eccl","Song","Isa","Jer","Lam","Ezek","Dan","Hos","Joel","Amos","Obad","Jonah","Mic","Nah","Hab","Zeph","Hag","Zech","Mal","Matt","Mark","Luke","John","Acts","Rom","1Cor","2Cor","Gal","Eph","Phil","Col","1Thess","2Thess","1Tim","2Tim","Titus","Phlm","Heb","Jas","1Pet","2Pet","1John","2John","3John","Jude","Rev"];var K=new c([{regEXP:/\[(.+?)\]/gi,elTag:"i"},{regEXP:/(LORD|God)/gi,elTag:"b"},{regEXP:/^(\d+)/gi,cls:"verseNumber"},{regEXP:/#/gi,cls:"versePBreak",replace:"\xB6"}]),Xe=new c([{regEXP:/\[(.+?)\]/gi,elTag:"i"},{regEXP:/(LORD|God)/gi,elTag:"b"},{regEXP:/^([^:]+:\d+)/gi,cls:"verseNumber"},{regEXP:/#/gi,cls:"versePBreak",replace:"\xB6"}]),p=class p{constructor(t,e,a){r(this,"book");r(this,"chapter");r(this,"verse");this.book=t,this.chapter=e,this.verse=a}text(t){return p.bible[t][this.book]?.[this.chapter]?.[this.verse]||""}crossRefs(){return(p.crossRefs[this.toOSIS()]||[]).map(e=>p.fromOSIS(e[0]))}toOSIS(){let t=p.booksOfTheBible.indexOf(this.book);return`${p.BookShortNames[t]||this.book}.${this.chapter}.${this.verse}`}static fromOSIS(t){let e=t.split("-"),[a,o,f]=e[0].split("."),d=p.BookShortNames.indexOf(a);if(d===-1)throw new Error(`Invalid book code: ${a}`);return new p(p.booksOfTheBible[d],parseInt(o),parseInt(f))}toString(){return`${this.book} ${this.chapter}:${this.verse}`}};r(p,"booksOfTheBible",V),r(p,"BookShortNames",G),r(p,"bible",{}),r(p,"crossRefs"),r(p,"topics");var s=p;var z={enableLogging:!0,debug:!1,reset:!1,includeAI:!1,flipPages:!1,style:{Foreground:"hsl(0, 100%, 100%)",Background:"hsl(0, 100%, 0%)",EnhanceSpacing:!0,Font:"Fontserif",fontSize:16},workspaces:{currentVerses:[new s("GENESIS",1,1),new s("JOHN",3,16),new s("PSALMS",23,2),new s("1 CORINTHIANS",13,4),new s("PHILIPPIANS",4,13),new s("ROMANS",8,28)]}};var D=class extends b{constructor(e,a){super(e,a);this.app=a;r(this,"_verse");this.ontitleclick(o=>{o.stopPropagation(),this.app.commandPalette.open(this.app.settings.workspaces.currentVerses)}),this.verse=new s("GENESIS",6,1)}get title(){return this.app.title}set title(e){this.app.title=e,this.titleEl&&(this.titleEl.textContent=e)}get verse(){return this._verse}set verse(e){this._verse=e,this.update()}update(){this.title=this._verse.toString().toTitleCase(),this.content.empty();let{book:e,chapter:a,verse:o}=this._verse;s.bible.KJV?.[e]?.[a].forEach((f,d)=>{if(d===0)return;let u=K.highlight(`${d} ${f}`),i=o===d,h=f.includes("#"),n=["verse"];i&&n.push("verseActive"),h&&n.push("versePBreak"),this.content.createEl("div",{text:u,cls:n},C=>{let I=new s(e,a,d);C.addEventListener("click",()=>{this.verse=I}),C.addEventListener("contextmenu",q=>{q.preventDefault(),q.stopPropagation(),this.app.commandPalette.open(I)})})}),this.content.scroll(0,0),this.content.querySelector(".verseActive")?.scrollIntoView({behavior:"smooth",block:"start"})}},v=class extends x{constructor(){super(...arguments);r(this,"verses",[]);r(this,"name","Verse List");r(this,"icon",M);r(this,"childtype",g.name)}onTrigger(e){Array.isArray(e)&&e[0]instanceof s?this.verses=e:this.verses=[]}getCommands(e){return this.getcompatible(e,this.verses,a=>a.toString(),a=>s.bible.KJV[a.book]?.[a.chapter]?.[a.verse]||"")}renderCommand(e,a){a.setTitle(e.toString().toTitleCase()).setDescription(s.bible.KJV[e.book]?.[e.chapter]?.[e.verse]||"").setSubsearch(!0)}executeCommand(e){this.app.MainScreen.verse=e,this.palette.close()}},g=class extends v{constructor(){super(...arguments);r(this,"name","Cross References");r(this,"icon",k)}onTrigger(e){e instanceof s?this.verses=e.crossRefs():e instanceof m?this.verses=e.verse.crossRefs():this.verses=[]}},m=class{constructor(t,e){this.specifity=t;this.verse=e}},R=class l extends x{constructor(){super(...arguments);r(this,"name","Go To Verse");r(this,"icon",A);r(this,"list",[]);r(this,"part",new m(0,new s("GENESIS",1,1)));r(this,"childtype",l.name)}onTrigger(e){if(e instanceof m)switch(this.part=e,e.specifity){case 0:this.name="Go To Verse",this.list=s.booksOfTheBible.map(a=>new m(1,new s(a,1,1)));break;case 1:this.name=`Go To Verse: ${e.verse.book}`,this.palette.inputMode="numeric",this.list=s.bible.KJV[e.verse.book]?.slice(1).map((a,o)=>new m(2,new s(e.verse.book,o+1,1)))||[];break;case 2:this.name=`Go To Verse: ${e.verse.book}:${e.verse.chapter}`,this.list=s.bible.KJV[e.verse.book]?.[e.verse.chapter]?.slice(1).map((a,o)=>new m(3,new s(e.verse.book,e.verse.chapter,o+1)))||[];break;case 3:this.app.MainScreen.verse=e.verse}else this.part=new m(0,new s("GENESIS",1,1)),this.list=s.booksOfTheBible.map(a=>new m(1,new s(a,1,1))),this.app.console.log("GoToVersePalette.onTrigger: No context provided, defaulting to books list",this.list)}getCommands(e){switch(this.part.specifity){case 0:return this.childtype=l.name,this.getcompatible(e,this.list,a=>a.verse.book);case 1:return this.childtype=l.name,this.getcompatible(e,this.list,a=>a.verse.chapter.toString());case 2:return this.childtype=g.name,this.getcompatible(e,this.list,a=>a.verse.verse.toString(),a=>s.bible.KJV[a.verse.book]?.[a.verse.chapter]?.[a.verse.verse]||"");default:return[]}}renderCommand(e,a){switch(this.part.specifity){case 0:a.setTitle(e.verse.book.toString().toTitleCase()).setSubsearch(!0);break;case 1:a.setTitle(`${e.verse.book.toString().toTitleCase()} ${e.verse.chapter}`).setSubsearch(!0);break;case 2:a.setTitle(e.verse.toString().toTitleCase()).setDescription(s.bible.KJV[e.verse.book]?.[e.verse.chapter]?.[e.verse.verse]||"");break}}executeCommand(e){e.specifity===3?(this.app.MainScreen.verse=e.verse,this.palette.close()):this.palette.display(e)}},H=class extends x{constructor(){super(...arguments);r(this,"name","Bible Search");r(this,"icon",L);r(this,"verses",[]);r(this,"bible",s.bible.KJV);r(this,"childtype",g.name)}onTrigger(e){this.app.console.log("BibleSearchPalette.onTrigger",e)}getCommands(e){let a=this.palette.maxResults-this.palette.commandItems.length;if(!e)return[];let o=[],f=e.toLowerCase();for(let d in this.bible)if(this.bible.hasOwnProperty(d)){let u=this.bible[d];for(let i=1;i<u.length;i++){let h=u[i];for(let n=1;n<h.length;n++){let C=h[n];if(C&&C.toLowerCase().includes(f)&&(o.push(new s(d,i,n)),o.length>a))return o}}}return o}renderCommand(e,a){a.setTitle(e.toString().toTitleCase()).setDescription(s.bible.KJV[e.book]?.[e.chapter]?.[e.verse]||"").setSubsearch(!0).setHidden(!1)}executeCommand(e){this.app.MainScreen.verse=e,this.palette.close()}},F=class l extends x{constructor(){super(...arguments);r(this,"topics",[]);r(this,"name","Topics");r(this,"icon",k);r(this,"childtype",l.name)}onTrigger(e){if(typeof e=="string"&&e.startsWith("topic:")){let a=e.slice(6);this.topics=s.topics[a].map(o=>s.fromOSIS(o[0])),this.name=`Topic: ${a.toTitleCase()}`}else this.topics=Object.keys(s.topics).map(a=>`topic:${a.toLowerCase()}`),this.name="Topics"}getCommands(e){return typeof this.topics[0]=="string"?e?(this.childtype=g.name,this.getcompatible(e,this.topics,a=>a.slice(6))):[]:(this.childtype=l.name,this.getcompatible(e,this.topics,a=>a.toString(),a=>s.bible.KJV[a.book]?.[a.chapter]?.[a.verse]||""))}renderCommand(e,a){typeof e=="string"?a.setTitle(e.slice(6).toTitleCase()).setSubsearch(!0):a.setTitle(e.toString().toTitleCase()).setDescription(s.bible.KJV[e.book]?.[e.chapter]?.[e.verse]||"").setSubsearch(!0)}executeCommand(e){typeof e=="string"?this.palette.display(e):(this.app.MainScreen.verse=e,this.palette.close())}},O=class extends y{constructor(e){super(e,"Touch Grass Bible");r(this,"settings");r(this,"commandPalette");r(this,"MainScreen")}async onload(){await this.loadsettings(z);let[e,a,o]=await Promise.all([this.loadJSON("KJV.json"),this.loadJSON("crossrefs.json"),this.loadJSON("topics.json")]);s.bible.KJV=e,s.crossRefs=a,s.topics=o,this.console.enabled=this.settings.enableLogging,this.console.header("color:#f0f; font-size:40px; font-weight:bold;"),this.MainScreen=new D(this.contentEl,this),this.commandPalette=new S(this),this.commandPalette.addPalette(v),this.commandPalette.addPalette(g),this.commandPalette.addPalette(F),this.commandPalette.addPalette(H),this.commandPalette.addPalette(R),this.console.log(new Date().getTime()-processstart,"ms startup time")}onunload(){return!0}onHistoryPop(e){e.name==="Command Palette"&&this.commandPalette.handleBack()}async loadsettings(e){this.settings=Object.assign({},e,await this.loadData())}saveSettings(){this.saveData(this.settings)}};var lt=new O(document);
-/*! Bundled license information:
-
-lucide/dist/esm/defaultAttributes.js:
-lucide/dist/esm/createElement.js:
-lucide/dist/esm/icons/book-copy.js:
-lucide/dist/esm/icons/chevron-right.js:
-lucide/dist/esm/icons/library.js:
-lucide/dist/esm/icons/list.js:
-lucide/dist/esm/icons/search.js:
-lucide/dist/esm/icons/x.js:
-lucide/dist/esm/lucide.js:
-  (**
-   * @license lucide v0.511.0 - ISC
-   *
-   * This source code is licensed under the ISC license.
-   * See the LICENSE file in the root directory of this source tree.
-   *)
-*/
