+++ processing/online.mjs
@@ -5,6 +5,7 @@ import { fileURLToPath } from "url";
 import { Open } from "unzipper";
 
 const _dirname = dirname(fileURLToPath(import.meta.url));
+const _dest = "./dest";
 const files = [
   {
     url: "https://a.openbible.info/data/topic-scores.zip",
@@ -19,12 +20,12 @@ const files = [
 function downloadZip(url, dest) {
   return new Promise((resolve, reject) => {
     const file = createWriteStream(dest);
-    get(url, (res) => {
+    get(url, res => {
       if (res.statusCode !== 200)
         return reject(new Error(`Failed to get '${url}' (${res.statusCode})`));
       res.pipe(file);
       file.on("finish", () => file.close(resolve));
-    }).on("error", (err) => reject(err));
+    }).on("error", err => reject(err));
   });
 }
 
@@ -37,57 +38,34 @@ async function unzipFirstFileAsString(zipPath) {
   try {
     for (const f of files) await downloadZip(f.url, f.path);
     const [topicsRaw, crossrefsRaw] = await Promise.all(
-      files.map((f) => unzipFirstFileAsString(f.path))
+      files.map(f => unzipFirstFileAsString(f.path))
     );
-    console.log("File 1:", topicsRaw.slice(0, 500));
-    console.log("File 2:", crossrefsRaw.slice(0, 500));
     const topics = {};
     topicsRaw
       .split("\n")
       .slice(1)
-      .forEach((line) => {
+      .forEach(line => {
         const [Topic, OSIS, QualityScore] = line.split("\t");
         if (!Topic || !OSIS || !QualityScore) return; // Skip invalid lines
         if (!topics[Topic]) topics[Topic] = [];
         topics[Topic].push([OSIS, Number(QualityScore)]);
       });
-    writeFileSync(join(_dirname, "topics.json"), JSON.stringify(topics));
-    writeFileSync(
-      join(_dirname, "topicsSample.json"),
-      JSON.stringify(
-        Object.fromEntries(
-          Object.keys(topics)
-            .slice(0, 10)
-            .map((k) => [k, topics[k]])
-        )
-      )
-    );
-    writeFileSync(join(_dirname, "topics.txt"), topicsRaw);
+    writeFileSync(join(_dest, "topics.json"), JSON.stringify(topics));
     const crossrefs = {};
     crossrefsRaw
       .split("\n")
       .slice(1)
-      .forEach((line) => {
+      .forEach(line => {
         const [FromVerse, ToVerse, Votes] = line.split("\t");
         if (!FromVerse || !ToVerse || !Votes) return; // Skip invalid lines
         if (!crossrefs[FromVerse]) crossrefs[FromVerse] = [];
         crossrefs[FromVerse].push([ToVerse, Number(Votes)]);
       });
-    writeFileSync(join(_dirname, "crossrefs.json"), JSON.stringify(crossrefs));
-    writeFileSync(
-      join(_dirname, "crossrefsSample.json"),
-      JSON.stringify(
-        Object.fromEntries(
-          Object.keys(crossrefs)
-            .slice(0, 10)
-            .map((k) => [k, crossrefs[k]])
-        )
-      )
-    );
-    writeFileSync(join(_dirname, "crossrefs.txt"), crossrefsRaw);
+    writeFileSync(join(_dest, "crossrefs.json"), JSON.stringify(crossrefs));
 
-    files.forEach((f) => unlinkSync(f.path));
+    files.forEach(f => unlinkSync(f.path));
     // Use data/data2 as needed
+    console.log("Data processing completed successfully.");
   } catch (err) {
     console.error("Error:", err);
   }
