+++ src/index.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="dest/main.css" />
+    <title>Real Bible App</title>
+  </head>
+  <body>
+    <script>
+      const processstart = new Date().getTime();
+    </script>
+    <script src="dest/main.js"></script>
+  </body>
+</html>
