--- webversion/index.html
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="main.css" />
-    <link rel="icon" href="./favicon.ico" />
-    <link rel="manifest" href="manifest.json" />
-    <link rel="icon" type="image/png" sizes="192x192" href="./icon-192.png" />
-    <link rel="icon" type="image/png" sizes="512x512" href="./icon-512.png" />
-    <title>Touch Grass Bible</title>
-  </head>
-  <body>
-    <script>
-      const processstart = new Date().getTime();
-    </script>
-    <script>
-      if ("serviceWorker" in navigator) {
-        navigator.serviceWorker
-          .register("./service-worker.js")
-          .then(registration => {
-            console.log("Service Worker registered with scope:", registration.scope);
-
-            fetch("version")
-              .then(response => response.text())
-              .then(version => (VERSION = version))
-              .catch(error => console.error("Error fetching version:", error));
-          })
-          .catch(error => {
-            console.error("Service Worker registration failed:", error);
-          });
-      }
-    </script>
-    <script src="main.js"></script>
-  </body>
-</html>
