--- webversion/manifest.json
@@ -1,22 +1,0 @@
-{
-  "name": "Touch Grass Bible",
-  "short_name": "TGBible",
-  "description": "A web version of the Touch Grass Bible app.",
-  "start_url": ".",
-  "display": "standalone",
-  "background_color": "#121212",
-  "theme_color": "#222b1a",
-  "icons": [
-    {
-      "src": "icons/icon-192.png",
-      "sizes": "192x192",
-      "type": "image/png"
-    },
-    {
-      "src": "icons/icon-512.png",
-      "sizes": "512x512",
-      "type": "image/png"
-    }
-  ],
-  "lang": "en"
-}
